refactor: вынесение AnnouncedPlayCardList из директории компонента AnnouncedPlayCard
