+++ src/components/ui/announced-play-card-list/announced-play-card-list.module.css
@@ -1,4 +1,4 @@
-.play-card-list {
+.list {
   display: flex;
   box-sizing: border-box;
   flex-wrap: wrap;
